+++ DustyPig.API/v3/Clients/FriendsClient.cs
@@ -1,5 +1,6 @@
 ﻿using DustyPig.API.v3.Models;
 using DustyPig.REST;
+using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,15 +18,25 @@ namespace DustyPig.API.v3.Clients
         /// <summary>
         /// Requires main profile
         /// </summary>
-        public Task<Response<DetailedFriend>> GetDetailsAsync(int id, CancellationToken cancellationToken = default) =>
-            _client.GetAsync<DetailedFriend>(true, PREFIX + $"Details/{id}", cancellationToken);
+        public Task<Response<DetailedFriend>> GetDetailsAsync(int id, CancellationToken cancellationToken = default)
+        {
+            if (id <= 0)
+                return Task.FromResult(new Response<DetailedFriend> { Error = new ModelValidationException($"Invalid {nameof(id)}") });
+
+            return _client.GetAsync<DetailedFriend>(true, PREFIX + $"Details/{id}", cancellationToken);
+        }
 
 
         /// <summary>
         /// Requires main profile. Invites a new friend using their email
         /// </summary>
-        public Task<Response> InviteAsync(string email, CancellationToken cancellationToken = default) =>
-            _client.PostAsync(true, PREFIX + "Invite", new SimpleValue<string>(email), cancellationToken);
+        public Task<Response> InviteAsync(string email, CancellationToken cancellationToken = default)
+        {
+            if(!StringUtils.IsValidEmail(email))
+                return Task.FromResult(new Response { Error = new ModelValidationException($"Invalid {nameof(email)}") });
+
+            return _client.PostAsync(true, PREFIX + "Invite", new SimpleValue<string>(email), cancellationToken);
+        }
 
 
         /// <summary>
@@ -45,8 +56,13 @@ namespace DustyPig.API.v3.Clients
         /// <summary>
         /// Requires main profile
         /// </summary>
-        public Task<Response> UnfriendAsync(int id, CancellationToken cancellationToken = default) =>
-            _client.DeleteAsync(true, PREFIX + $"Unfriend/{id}", cancellationToken);
+        public Task<Response> UnfriendAsync(int id, CancellationToken cancellationToken = default)
+        {
+            if (id <= 0)
+                return Task.FromResult(new Response { Error = new ModelValidationException($"Invalid {nameof(id)}") });
+
+            return _client.DeleteAsync(true, PREFIX + $"Unfriend/{id}", cancellationToken);
+        }
 
 
         /// <summary>
