+++ DustyPig.API/v3/Clients/NotificationsClient.cs
@@ -18,8 +18,13 @@ namespace DustyPig.API.v3.Clients
         /// <summary>
         /// Requires profile
         /// </summary>
-        public Task<Response> DeleteAsync(int id, CancellationToken cancellationToken = default) =>
-            _client.DeleteAsync(true, PREFIX + $"Delete/{id}", cancellationToken);
+        public Task<Response> DeleteAsync(int id, CancellationToken cancellationToken = default)
+        {
+            if (id <= 0)
+                return Task.FromResult(new Response { Error = new ModelValidationException($"Invalid {nameof(id)}") });
+        
+            return _client.DeleteAsync(true, PREFIX + $"Delete/{id}", cancellationToken);
+        }
 
 
         /// <summary>
@@ -32,8 +37,13 @@ namespace DustyPig.API.v3.Clients
         /// <summary>
         /// Requires profile
         /// </summary>
-        public Task<Response> MarkAsReadAsync(int id, CancellationToken cancellationToken = default) =>
-            _client.GetAsync(true, PREFIX + $"MarkAsRead/{id}", cancellationToken);
+        public Task<Response> MarkAsReadAsync(int id, CancellationToken cancellationToken = default) 
+        {
+            if (id <= 0)
+                return Task.FromResult(new Response { Error = new ModelValidationException($"Invalid {nameof(id)}") });
+        
+            return _client.GetAsync(true, PREFIX + $"MarkAsRead/{id}", cancellationToken);
+        }
 
     }
 }
